+++ plugins/postcss-normalize-display-values/CHANGELOG.md
@@ -1,17 +1,25 @@
 # Changes to PostCSS Normalize Display Values
 
-### 2.0.1 (January 28, 2023)
+### 2.0.1
+
+_January 28, 2023_
 
 - Improve `types` declaration in `package.json`
 
-### 2.0.0 (January 24, 2023)
+### 2.0.0
+
+_January 24, 2023_
 
 - Updated: Support for Node v14+ (major).
 
-### 1.0.1 (July 8, 2022)
+### 1.0.1
+
+_July 8, 2022_
 
 - Fix case insensitive matching.
 
-### 1.0.0 (January 25, 2022)
+### 1.0.0
+
+_January 25, 2022_
 
 - Initial version